backend/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
google-generativeai
pydantic
langchain # For document loading and chunking, if not implemented manually
python-multipart # Needed for file uploads, if any
